Initial plugin featuring `.search`, `.suggest`, & `.gsuggest` commands
`.gsuggest` requires extra dependencies and is optional (using the very
creatively named 'gsuggest' setuptools extra).

`.search` and `.suggest` use the `duckduckgo-search` library, and obey
a couple of simple configuration values (`region` and `safesearch`).

Tested in Gitpod development environment, but not yet "in production" on
a Sopel bot where the built-in `search.py` mysteriously fails to find
any results; that's the next step before release, to see if this is even
an improvement vs. the status quo.