Also immediately refresh DDGS client after exceptions
The client object remembers if it encountered an error, and will bail on
further requests immediately with:

    DuckDuckGoSearchException("Exception occurred in previous call.")

If one of the recognizable exception types happens, the client therefore
has to be replaced before the user retries.

The patch doesn't apply this to other exception types, as it isn't clear
whether that would even help. (If it turns out that it would help,
that's what updates are for.)