Specify a list of backends to try
In testing with DDGS 9.6, the result was garbage a lot of the time:

```
<dgw> ,search sopel bot
<SopelTest> [search] Solel Boneh —
            https://en.wikipedia.org/wiki/Solel_Boneh
<dgw> ,search sopel irc
<SopelTest> [search] Sope Dirisu —
            https://en.wikipedia.org/wiki/Sope_Dirisu
<dgw> ,search sopel irc bot
<SopelTest> [search] Cm to Feet +Inches Converter ( cm to ft) - The
            Calculator Site —
            https://www.thecalculatorsite.com/conversions/common/cm-to-feet-inches.php
<dgw> what in the fuck is this garbage
```

In part, this was because "auto" places Wikipedia at the beginning of
the priority list no matter what. But the full set of available search
engines is also just..... too much.

This hard-codes a preference for DDG, Google, Brave, and Bing (in that
order). Maybe it should be configurable, but let's try it out first.
(Getting the list of available engines and presenting it in the config
wizard might be a bit of a challenge. DDGS says that user code shouldn't
import its `ENGINES` dict.)